+++ tests/clock_tests.c
@@ -3,25 +3,25 @@
 
 #include "clock.h"
 
-static const long long int SLEPS = 1;
+static const uint64_t SLEPS = 1;
 
 int main(int argc, char *argv[])
 {
-        printf("clkfreq  =%lld\n", clkfreq());
-        printf("clkelapt =%lld\n", clkelapt());
-        printf("clkelaps =%lld\n", clkelaps());
-        printf("clkelapms=%lld\n", clkelapms());
-        printf("clkelapus=%lld\n", clkelapus());
-        printf("clkelapns=%lld\n", clkelapns());
-        printf("clkslept (%lld)...\n", SLEPS * clkfreq());
+        printf("clkfreq  =%lu\n", clkfreq());
+        printf("clkelapt =%lu\n", clkelapt());
+        printf("clkelaps =%lu\n", clkelaps());
+        printf("clkelapms=%lu\n", clkelapms());
+        printf("clkelapus=%lu\n", clkelapus());
+        printf("clkelapns=%lu\n", clkelapns());
+        printf("clkslept (%lu)...\n", SLEPS * clkfreq());
         clkslept(SLEPS * clkfreq());
-        printf("clksleps (%lld)...\n", SLEPS);
+        printf("clksleps (%lu)...\n", SLEPS);
         clksleps(SLEPS);
-        printf("clkslepms(%lld)...\n", SLEPS * MSPERS);
+        printf("clkslepms(%lu)...\n", SLEPS * MSPERS);
         clkslepms(SLEPS * MSPERS);
-        printf("clkslepus(%lld)...\n", SLEPS * USPERS);
+        printf("clkslepus(%lu)...\n", SLEPS * USPERS);
         clkslepus(SLEPS * USPERS);
-        printf("clkslepns(%lld)...\n", SLEPS * NSPERS);
+        printf("clkslepns(%lu)...\n", SLEPS * NSPERS);
         clkslepns(SLEPS * NSPERS);
         return EXIT_SUCCESS;
 }
